refactor(sidecar): extract pure planner.py from accel.py
Move deployment-planning logic (resolve/resolve_translate/resolve_tts, tier
gating, quant/variant pickers) into sokuji_sidecar/planner.py, leaving
accel.py as the Loader (hardware probe, download, load, measure, RPC
handlers). accel.py re-exports the moved names so accel.<name> keeps
resolving for the frozen characterisation suite and existing callers; the
three engines now call planner.resolve*() directly.

planner.py reaches a handful of accel-owned, test-monkeypatched environment
helpers (probe, current_platform, _downloaded_quants, bench_load,
_format_ready, _est_bytes) via deferred `from . import accel` inside the
functions that need them, rather than importing them by value, so existing
monkeypatch.setattr(accel, ...) patches in the test suite keep taking
effect after the split.

Also updates test_asr_engine.py/test_translate_engine.py/test_tts_engine.py,
whose resolver monkeypatches targeted accel.resolve* and must now target
planner.resolve* to match the engines' new call sites.
