+++ sidecar/tests/test_tts_engine.py
@@ -1,7 +1,7 @@
 import asyncio
 import numpy as np
 import pytest
-from sokuji_sidecar import tts_engine, accel, catalog
+from sokuji_sidecar import tts_engine, accel, catalog, planner
 
 
 def test_resample_48k_stereo_to_24k_mono():
@@ -62,7 +62,7 @@ class _FakeStream:
 
 def _patch(monkeypatch, backend, model_id):
     plan = accel.Plan(backend.NAME, "cpu", "cpu", "fp32", "repo", 1.0)
-    monkeypatch.setattr(accel, "resolve_tts", lambda *a, **k: [plan])
+    monkeypatch.setattr(planner, "resolve_tts", lambda *a, **k: [plan])
     monkeypatch.setattr(accel, "load_measured", lambda plans, **kw: (backend, plan, None, None))
     monkeypatch.setattr(accel, "measure_rtf_tts", lambda *a, **k: 0.1)
